Add MaxAttempts to Backoff
The field MaxAttempts is added to the Backoff struct, this allows
the Duration method to fail with an error if a number of attempts
have been exceeded. The signature of Duration has been changed to
return a duration and an error.

Also this commit changes attempts from float64 to int and creates
a new test when exceeding the number of attempts.

Fixes: #2